Migrates the types of length and width from int to double

+++ src/com/tw/bootcamp/problem1/Rectangle.java
@@ -2,27 +2,27 @@ package com.tw.bootcamp.problem1;
 
 public class Rectangle {
 
-  private final int width;
-  private final int length;
+  private final double width;
+  private final double length;
 
-  private Rectangle(int width, int length) {
+  private Rectangle(double width, double length) {
     this.width = width;
     this.length = length;
   }
 
-  public static Rectangle createRectangle(int width, int length) {
+  public static Rectangle createRectangle(double width, double length) {
     return new Rectangle(width, length);
   }
 
-  public static Rectangle createSquare(int side) {
+  public static Rectangle createSquare(double side) {
     return new Rectangle(side , side);
   }
 
-  public int area() {
+  public double area() {
     return width * length;
   }
 
-  public int perimeter() {
+  public double perimeter() {
     return 2 * (width + length);
   }
 }
